+++ src/scripts/store.js
@@ -21,24 +21,6 @@ const store = createStore({
     },
   },
   actions: {
-    login({commit}, userData) {
-      axios.post("/users/login", userData, {withCredentials: true})
-      .then((res) => {
-        const token = res.headers['authorization'];
-        axios.defaults.headers.common['Authorization'] = token;
-        commit('setToken', token);
-        const accessToken = {
-          token: token,
-          expire: Date.now() + (60 * 60 * 1000),
-        }
-        localStorage.setItem("accessToken", JSON.stringify(accessToken));
-        location.reload();
-      })
-      .catch(() => {
-        alert("로그인 실패");
-      })
-    },
-
     logout({commit}) {
       alert('로그아웃 성공');
       localStorage.removeItem("accessToken");
